[refactor] RawHttpResponse를 /dto/rawHttp로 이동 #30
