+++ src/main/kotlin/dev/codeswamp/global/auth/application/dto/rawHttp/RawHttpResponse.kt
@@ -1,4 +1,4 @@
-package dev.codeswamp.global.auth.application.dto
+package dev.codeswamp.global.auth.application.dto.rawHttp
 
 data class RawHttpResponse (
     val cookies: MutableMap<String, String>
